FeatureStylesService: move from /directives to /services folder
