+++ projects/core/src/features-config/index.ts
@@ -9,5 +9,6 @@ export * from './directives/feature-level.directive';
 export * from './directives/feature.directive';
 export * from './features-config.module';
 export * from './services/feature-config.service';
+export * from './services/feature-styles.service';
 export * from './utils/feature-config-utils';
 export * from './utils/use-feature-styles';
